requirements.txt
datasets>=2.18.0
huggingface-hub>=0.22.0
pandas>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
groq>=0.8.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
